cmb(3): API changes to cmb_print(1) and struct cmb_config
Change the number of arguments that are passed to the action function
(default being cmb_print()). Introduce new first argument (changing
the total number of arguments from 2 to 3), struct cmb_config *config.

This allows me to eliminate the static globals in cmb.c to prevent
race-conditions in multi-threaded applications. It also cleans up the
processing of the config struct while improving accuracy of the manual
(which talks of cmb_print()'s direct use of config, not globals which
were only updated by a call to either cmb() or cmb_bn()).

Last, but not least, this opens the ability to use cmb_print() without
first calling cmb() or cmb_bn().